LoadSystem-async-service/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
requests==2.32.5

